fixed itinerary class - removed customer. fixed ticket class. Added souts to main for ckecks. created some customers itineraries and tickets.

+++ src/com/travelcompany/eshop/service/CustomerService.java
@@ -10,5 +10,5 @@ public interface CustomerService {
 
     void addCustomer(Customer customer);
 
-    Customer getCustomerById(int id);
+    Customer getCustomerById(long id);
 }
