+++ src/com/travelcompany/eshop/service/CustomerServiceImpl.java
@@ -21,7 +21,7 @@ public class CustomerServiceImpl implements CustomerService {
     }
 
     @Override
-    public Customer getCustomerById(int id) {
+    public Customer getCustomerById(long id) {
         for (Customer customer : customerList) {
             if (customer.getId() == id) {
                 return customer;
@@ -29,4 +29,10 @@ public class CustomerServiceImpl implements CustomerService {
         }
         return null;
     }
+
+    //πρέπει να κάνω δική μου λίστα που να ξεκινά το id απο ένα μέχρι και αυτή τη λίστα να βάλω στο = new Arraylist<customer>();
+    //να βάλω ενα κανόνα που θα δεν θα επιτρέπει τον ίδιο customer 2 φορές.
+    // Set<Customer> customer = new Hashset<Customer>(); μια λίστα που επιτρέπει μια φορά να καταχωρηθεί το καθένα
+    // το set δεν έχει τη δυνατότητα να σου φέρει τη θέση του κάθε πελάτη. Αλλά είναι πολύ γρήγορα στην αναζήτηση
+
 }
